refactor: 게시물 라우터 주소 변경 - posts -> blogs
