+++ constants/routes.ts
@@ -1,7 +1,7 @@
 export const ROUTES = {
   HOME: '/',
   PAGES: '/pages',
-  POSTS: '/posts',
+  BLOGS: '/blogs',
   TAGS: '/tags',
   ABOUT: '/about',
 } as const;
