+++ utils/createPostPath.ts
@@ -2,7 +2,7 @@ import { FILE_NAME } from '../constants/blogPosts';
 import { ROUTES } from '../constants/routes';
 
 const createPostPath = (slug: FILE_NAME) => {
-  return ROUTES.POSTS + `/${slug}`;
+  return ROUTES.BLOGS + `/${slug}`;
 };
 
 export default createPostPath;
